real_estate_voiceai/requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
vapi-python
